+++ src/Toolkit/Toolkit/UI/Controls/OfflineMapAreas/OfflineMapViewModel.cs
@@ -409,68 +409,4 @@ namespace Esri.ArcGISRuntime.Toolkit.UI.Controls
             return parent?.Name;
         }
     }
-
-    /// <summary>
-    /// Set of helper utility classes for the Offline Map Areas related classes.
-    /// </summary>
-    internal static class OfflineMapAreaUtilities
-    {
-        public static long GetDirectorySize(string directory)
-        {
-            if (!Directory.Exists(directory))
-            {
-                return 0;
-            }
-
-            return Directory
-                .EnumerateFiles(directory, "*", SearchOption.AllDirectories)
-                .Select(file => new FileInfo(file).Length)
-                .Sum();
-        }
-
-        public static bool IsNetworkAvailable() => NetworkInterface.GetIsNetworkAvailable();
-
-        public static bool IsCancellation(IJob job, Exception ex)
-            => ex is OperationCanceledException || ex is TaskCanceledException || job.Status == JobStatus.Canceling;
-
-        public static async Task<byte[]?> LoadImageBytesAsync(RuntimeImage? image)
-        {
-            if (image is null)
-            {
-                return null;
-            }
-
-            if (image.LoadStatus != LoadStatus.Loaded)
-            {
-                await image.LoadAsync().ConfigureAwait(false);
-            }
-
-            if (image.LoadStatus != LoadStatus.Loaded)
-            {
-                return null;
-            }
-
-            using var stream = await image.GetEncodedBufferAsync().ConfigureAwait(false);
-            var bytes = new byte[stream.Length];
-            await stream.ReadExactlyAsync(bytes).ConfigureAwait(false);
-            return bytes;
-        }
-
-        public static void TryDeleteDirectory(string directory)
-        {
-            if (!Directory.Exists(directory))
-            {
-                return;
-            }
-
-            try
-            {
-                Directory.Delete(directory, recursive: true);
-            }
-            catch
-            {
-                // Best effort cleanup only.
-            }
-        }
-    }
 }
